Cleaned up generate wallet to be many different components. Added a paper wallet / private key step thats all fake right now.
