+++ common/containers/Tabs/GenerateWallet/components/PasswordInput.jsx
