+++ common/containers/Tabs/GenerateWallet/index.js
@@ -1,8 +1,10 @@
-import GenerateWalletPasswordComponent from './components/GenerateWalletPasswordComponent';
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import * as generateWalletActions from 'actions/generateWallet';
 import PropTypes from 'prop-types';
+import EnterPassword from './components/EnterPassword';
+import SaveWallet from './components/SaveWallet';
+import PaperWallet from './components/PaperWallet';
 
 class GenerateWallet extends Component {
   constructor(props) {
@@ -11,8 +13,8 @@ class GenerateWallet extends Component {
 
   static propTypes = {
     // state
+    activeStep: PropTypes.string,
     generateWalletPassword: PropTypes.object,
-    showPassword: PropTypes.bool,
     hasDownloadedWalletFile: PropTypes.bool,
     generateWalletFile: PropTypes.bool,
     canProceedToPaper: PropTypes.bool,
@@ -24,15 +26,46 @@ class GenerateWallet extends Component {
   };
 
   render() {
-    return <GenerateWalletPasswordComponent {...this.props} />;
+    const { activeStep } = this.props;
+    let content;
+
+    switch (activeStep) {
+      case 'password':
+        content = <EnterPassword {...this.props} />;
+        break;
+
+      case 'download':
+        content = <SaveWallet {...this.props} />;
+        break;
+
+      case 'paper':
+        content = <PaperWallet {...this.props} />;
+        break;
+
+      default:
+        content = <h1>Uh oh. Not sure how you got here.</h1>;
+    }
+
+    return (
+      <section className="container" style={{ minHeight: '50%' }}>
+        <div className="tab-content">
+          <main className="tab-pane active text-center" role="main">
+            <section role="main" className="row">
+              <br />
+              {content}
+            </section>
+          </main>
+        </div>
+      </section>
+    );
   }
 }
 
 function mapStateToProps(state) {
   return {
     generateWalletPassword: state.form.generateWalletPassword,
+    activeStep: state.generateWallet.activeStep,
     generateWalletFile: state.generateWallet.generateWalletFile,
-    showPassword: state.generateWallet.showPassword,
     hasDownloadedWalletFile: state.generateWallet.hasDownloadedWalletFile,
     canProceedToPaper: state.generateWallet.canProceedToPaper
   };
